requirements-dev.txt
build~=0.10.0
tomli~=2.0.1
pytest-cov~=4.0.0
sphinx~=6.1.3
pylint~=2.17.0
twine~=4.0.2
